refactor: move `src/optimizer/utils` to `src/ast/`
Fixes #1421
