+++ src/optimizer/standardOptimizer.ts
@@ -23,7 +23,7 @@ import {
     AssociativeRule3,
 } from "./associative";
 import { Rule, ExpressionTransformer } from "./types";
-import { AstUtil } from "./util";
+import { AstUtil } from "../ast/util";
 
 type PrioritizedRule = { priority: number; rule: Rule };
